[F] Sort DSQ status (#228)
* Add Sort option for status DSQ

* Fix tests

* Add test for default sort option

+++ test/iRLeagueApiCore.Services.Tests/ResultService/Calculation/TeamSessionCalculationServiceTests.cs
@@ -212,7 +212,7 @@ public sealed class TeamSessionCalculationServiceTests
                 testRow.FinalPosition.Should().Be(position + penalty.Positions);
                 break;
             case PenaltyType.Disqualification:
-                testRow.Status.Should().Be((int)RaceStatus.Disqualified);
+                testRow.Status.Should().Be(RaceStatus.Disqualified);
                 break;
             default:
                 break;
